Add container style prop
Currently there is no way to style the Ticker container, say if one wants to add a margin to the left or right of the Ticker without needing to wrap the ticker in another View it would be better to pass a style prop. 

Usually I would add the margin to the textStyle prop but since this components text / children is mapped to individual text components textStyle is applied to each meaning the margin is applied to each Text component.